ingest/fx/historical/requirements.txt
google-cloud-bigquery==3.17.2
functions-framework==3.4.0
requests==2.31.0